common: stub to use user-defined rules
The rules used to perform pattern-matching are specified by the user,
and they should be fetched from the configuration of the addon.
Things fetched from the addon's config are resolved with a Promise.
Therefore there is lot of code that must be updated in order to be able
to work with the Promise.