common: use incremental number to identify rules
We have no constraints on the name or regex of rules, this means that we
can end up having duplicated names and/or regex.
This is a problem for those operations that must manage a specific rule.
Instead of adding consistency constraints on rule's name/regex, we just
add a simple incremental id.